% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_join.R
\name{ext_join}
\alias{ext_join}
\title{ext_join Function}
\usage{
ext_join(df)
}
\description{
This function allows you to join data from an external scheme
to the integrated species dictionary.
}
\examples{
ext_join()
}
\keyword{plants}
