% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_avg.R
\name{ind_avg}
\alias{ind_avg}
\title{ind_avg Function}
\usage{
ind_avg(df)
}
\description{
This function allows you to calculate the average indicator value for each
plot. Currently includes Ellenberg N, R and W.
}
\examples{
ind_avg()
}
\keyword{plants}
