% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_cwt.R
\name{ind_cwt}
\alias{ind_cwt}
\title{ind_cwt Function}
\usage{
ind_cwt(df)
}
\description{
This function allows you to calculate the cover-weighted indicator values for
each plot. Currently includes Ellenberg N, R and W.
}
\examples{
ind_cwt()
}
\keyword{plants}
