% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_join.R
\name{scheme_join}
\alias{scheme_join}
\title{scheme_join Function}
\usage{
scheme_join(df)
}
\description{
This function allows you to join data from the CS, ECN, LTMN or NPMS
to the integrated species dictionary.
}
\examples{
scheme_join()
}
\keyword{plants}
